fixed pest file color map, forced change for pst.
If a preset file had a non-standard color-map, it was screwing with
LEDs. So CMD 4 gets sent at the end of the dump routine. I also forced
a change to the file input so you can change banks and load the same
preset file into each bank. Finally, I forced the current bank into the
SX messages that come from a preset file, since that would always
update with the current preset.